Crate oppgave

oppgave - A simple Redis-based task queue

oppgave provides a small reliable queue on top of Redis. It allows to push tasks and fetch them again.

Tasks can be arbitrary objects, as long as they can be encoded and decoded into a String. The easiest way is to rely on JSON encoding by marking the task as RustcEncodable and RustcDecodable.

Oppgave prodives a reliable queue by moving acquired tasks to a backup queue. If a task finished it is removed from this backup queue. If a task fails it remains in the backup queue for human processing later on.

See Queue for a detailed documentation how to use this.

The following examples are provided as executables as well:

Example: Producer

#[derive(RustcDecodable, RustcEncodable)]
struct Job { id: u64 }

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();
let producer = Queue::new("default".into(), con);

producer.push(Job{ id: 42 });

Example: Worker

#[derive(RustcDecodable, RustcEncodable)]
struct Job { id: u64 }

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();
let worker = Queue::new("default".into(), con);

while let Some(task) = worker.next() {
    println!("Working with Job {}", job.id);
}

Structs

Queue
A Queue allows to push new tasks or fetch and decode them for processing.

TaskGuard
A wrapper of the fetched task.

Traits

TaskDecodable
Task objects that can be reconstructed from the data stored in Redis

TaskEncodable
Task objects that can be encoded to a string to be stored in Redis