Struct oppgave::Queue

pub struct Queue {
    // some fields omitted
}

A Queue allows to push new tasks or fetch and decode them for processing.

Push

Pushing new tasks to the queue encodes the given object and stores it in Redis for later processing.

Example

#[derive(RustcDecodable, RustcEncodable)]
struct Job { id: u64 }

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();
let producer = Queue::new("default".into(), con);

producer.push(Job{ id: 42 });

Fetch

A Queue provides a convenient Iterator-like interface over tasks:

#[derive(RustcDecodable, RustcEncodable)]
struct Job { id: u64 }

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();
let queue = Queue::new("default".into(), con);

while let Some(task) = queue.next() {
    println!("Working with Job {}", job.id);
}

Fetching a task from a queue returns a wrapper object, which delegates to the underlying, automatically decoded task object. If this wrapper object is dropped, the task is considered complete and therefore removed from the queue. If the task processing fails, you need to call fail() on this wrapper.

Example: Task complete & task failed

struct Job { id: u64 }

let client = redis::Client::open("redis://127.0.0.1/").unwrap();
let con = client.get_connection().unwrap();
let worker = Queue::new("default".into(), con);

{
  // `next` gives an Option<Result<...>>
  let task = worker.next().unwrap().unwrap();
} // Task succeeded, removed from backup queue

{
  let task = worker.next().unwrap().unwrap();
  task.fail();
} // Task failed, stays in backup queue

Methods

impl Queue

fn new(name: String, client: Connection) -> Queue

Create a new Queue for the given name

fn stop(&self)

Stop processing the queue

On the next .next() call None will be returned.

fn is_stopped(&self) -> bool

Check if queue processing is stopped

fn queue(&self) -> &str

Get the full queue name

fn backup_queue(&self) -> &str

Get the full backup queue name

fn size(&self) -> u64

Get the number of remaining tasks in the queue

fn push<T: TaskEncodable>(&self, task: T) -> RedisResult<()>

Push a new task to the queue

fn next<T: TaskDecodable>(&self) -> Option<RedisResult<TaskGuard<T>>>

Grab the next task from the queue

This method blocks and waits until a new task is available.